Retype EIP to u64

+++ src/cpu.rs
@@ -3,7 +3,7 @@ use std::sync::{mpsc::Sender, Arc, Barrier};
 use crate::msg::UIMessage;
 
 pub fn cpu_loop(mem: &mut [u8], cpu_barrier: Arc<Barrier>, ui_sender: Sender<UIMessage>) {
-    let mut eip: i64 = 0;
+    let mut eip: u64 = 0;
     loop {
         cpu_barrier.wait();
         for _i in 0..1000 {
@@ -37,7 +37,7 @@ pub fn cpu_loop(mem: &mut [u8], cpu_barrier: Arc<Barrier>, ui_sender: Sender<UIM
             a_val = a_val.wrapping_sub(b_val);
             crate::mem::write(mem, a_addr as usize, &i64::to_be_bytes(a_val));
             if a_val <= 0 {
-                eip = c_addr;
+                eip = c_addr as u64;
             } else {
                 eip += 24;
             }
